pyproject.toml
[tool.poetry]
name = "foodlabel"
version = "0.1.0"
description = ""
authors = ["codeblech <my.malikyash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.24.0"
google-generativeai = "^0.8.1"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
flask = "^3.0.3"
flask-cors = "^5.0.0"
werkzeug = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
